[refactor] Make rule enablement extensible.
This PR adds an extension hook for disabling rules.
This will allow us to disable rules with visibility labels internally.
